2010년 11월 5일 master와 머지.

+++ source/OliveLib/include/OvMesh.h
@@ -4,6 +4,10 @@
 #include "OvRenderingCommon.h"
 #include "OvBitFlags.h"
 
+
+// 메쉬 포맷을 3단계로 고정시키자
+// 유연해봐야 쓰지도 않는다.
+
 OvREF_POINTER(OvMesh)
 class OvMesh : public OvResource
 {
@@ -22,9 +26,9 @@ public:
 		SRenderData():faceStream(NULL), vertDecl(NULL), vertexCount(0), faceCount(0){};
 		SVertexStreamInfo	vertStreamInfo[ STAGE_MAX ];
 		LPDIRECT3DVERTEXDECLARATION9 vertDecl;
-		size_t	vertexCount;
+		UINT	vertexCount;
 		LPDIRECT3DINDEXBUFFER9	faceStream;
-		size_t	faceCount;
+		UINT	faceCount;
 	};
 
 	OvMesh();
