+++ source/OliveLib/source/OvMeshLoaderBase.cpp
@@ -1,4 +1,5 @@
 #include "OvMeshLoaderBase.h"
+#include "OvMesh.h"
 #include "OvAutoPtr.h"
 
 OvRTTI_IMPL(OvMeshLoaderBase);
@@ -15,9 +16,9 @@ OvMeshSPtr OvMeshLoaderBase::_buildMesh( SVertexStreamInfo& streamLow
 								 , SVertexStreamInfo& streamMedium
 								 , SVertexStreamInfo& streamHigh
 								 , LPDIRECT3DVERTEXDECLARATION9 declVert
-								 , size_t countVert
+								 , UINT countVert
 								 , LPDIRECT3DINDEXBUFFER9 streamFace
-								 , size_t countFace )
+								 , UINT countFace )
 {
 	OvAutoPtr< OvMesh::SRenderData > renderData = OvNew OvMesh::SRenderData;
